refactor: extract services into domain
