+++ platform/api/hexgate_api/bootstrap/__init__.py
@@ -1,0 +1,1 @@
+"""Startup-time seeding + first-boot bootstrap."""
