+++ platform/api/hexgate_api/domains/agents/service.py
@@ -1,444 +1,26 @@
+"""Agent + version persistence, policy-bundle compilation, manifest registration.
+
+Groups the agent read/write helpers, the save-time WASM bundle compile+sign
+(``compile_bundle`` shells out to the SDK/opa), and the ``hexgate register``
+upsert path (manifest → Agent + AgentVersion + Tool rows, with a generated
+starter policy on first registration).
+"""
+
 import hashlib
 import json
 import logging
-import os
-import secrets
 from typing import Callable
 
 from sqlalchemy import func
 from sqlmodel import select
 from sqlmodel.ext.asyncio.session import AsyncSession
 
-from hexgate_api.models import (
-    Agent,
-    AgentVersion,
-    Invitation,
-    Organization,
-    OrganizationMember,
-    Project,
-    Tool,
-    User,
-    utcnow,
-)
-from hexgate_api.constants import (
-    ALL_ROLES,
-    DEFAULT_MEMBERSHIP_ID,
-    DEFAULT_ORG_ID,
-    DEFAULT_ORG_NAME,
-    DEFAULT_ORG_SLUG,
-    DEFAULT_PROJECT_ID,
-    DEFAULT_PROJECT_NAME,
-    DEFAULT_USER_EMAIL,
-    DEFAULT_USER_ID,
-)
 from hexgate_api.core.ids import new_id
-from hexgate_api.domains.members.service import (
-    _ROLE_RANK,
-    _can_invite_role,
-    find_member,
-)
+from hexgate_api.models import Agent, AgentVersion, Tool
 from hexgate_api.schemas import AgentManifest, ToolDefinition
 from hexgate_api.seeds import SEED_AGENTS
 
-logger = logging.getLogger("hexgate.platform.services")
-
-
-def _seed_disabled() -> bool:
-    """``HEXGATE_SEED=skip`` opts a deployment out of the triple-default."""
-    return os.environ.get("HEXGATE_SEED", "").strip().lower() == "skip"
-
-
-# ---------------------------------------------------------------------------
-# M3 Phase 4 step 4 — Invitations
-#
-# Admins/owners mint an Invitation row; the row's id doubles as the
-# unguessable magic-link token. The invitee clicks an emailed
-# ``${dashboard}/invites/{id}/accept`` link, the dashboard logs them in
-# (registering first if needed) and POSTs /accept which atomically
-# creates the OrganizationMember row + marks the invite consumed.
-# ---------------------------------------------------------------------------
-
-
-# Default lifetime — a week is enough for "I'll get to it on Monday" but
-# short enough that an exfiltrated email link 6 months later is dead.
-_DEFAULT_INVITE_TTL_HOURS = 168
-
-
-class InvitationError(Exception):
-    """Base class for invitation-validation failures.
-
-    Routes translate the subclasses to specific HTTP codes:
-      * :class:`InvitationExpired`         → 410 Gone
-      * :class:`InvitationAlreadyConsumed` → 409 Conflict
-      * :class:`InvitationEmailMismatch`   → 403 Forbidden
-      * any other InvitationError          → 400 Bad Request
-    """
-
-
-class InvitationExpired(InvitationError):
-    """The ``expires_at`` deadline has passed."""
-
-
-class InvitationAlreadyConsumed(InvitationError):
-    """The invitation was already accepted or revoked — terminal."""
-
-
-class InvitationEmailMismatch(InvitationError):
-    """The signed-in user's email doesn't match the invited address.
-
-    The invite is for a specific person, not a bearer-token-for-anyone.
-    Refuse strictly rather than allow anyone-with-the-link to join."""
-
-
-async def create_invitation(
-    session: AsyncSession,
-    *,
-    org_id: str,
-    email: str,
-    role: str,
-    invited_by: OrganizationMember,
-    ttl_hours: int = _DEFAULT_INVITE_TTL_HOURS,
-) -> Invitation:
-    """Mint a pending invitation. Cancels any existing pending invite
-    for the same (org, email) pair so a re-invite produces one
-    valid link rather than two.
-
-    Refuses with ``InvitationError`` if the role is unknown or if the
-    inviter doesn't outrank the target role (see :func:`_can_invite_role`).
-    Email is lowercased so the case-insensitive accept-match works
-    even on SQLite (no native CI collations).
-    """
-    from datetime import timedelta
-
-    if role not in ALL_ROLES:
-        raise InvitationError(f"unknown role: {role!r}")
-    if not _can_invite_role(invited_by.role, role):
-        raise InvitationError(f"a {invited_by.role} cannot invite at the {role} level")
-
-    normalized_email = email.strip().lower()
-
-    # Cancel any existing pending invite for this (org, email). Two
-    # parallel invites would let either link redeem the seat — we want
-    # exactly one canonical pending link per address per org.
-    existing = (
-        await session.exec(
-            select(Invitation).where(
-                Invitation.org_id == org_id,
-                Invitation.email == normalized_email,
-                Invitation.accepted_at.is_(None),  # type: ignore[union-attr]
-                Invitation.revoked_at.is_(None),  # type: ignore[union-attr]
-            )
-        )
-    ).all()
-    for stale in existing:
-        stale.revoked_at = utcnow()
-        session.add(stale)
-
-    invitation = Invitation(
-        org_id=org_id,
-        email=normalized_email,
-        role=role,
-        invited_by_user_id=invited_by.user_id,
-        expires_at=utcnow() + timedelta(hours=ttl_hours),
-    )
-    session.add(invitation)
-    await session.commit()
-    await session.refresh(invitation)
-    return invitation
-
-
-async def find_invitation(
-    session: AsyncSession, invitation_id: str
-) -> Invitation | None:
-    """Return the Invitation by id, or None. Doesn't validate
-    expiry / consumed state — the caller does that (the preview
-    endpoint shows expired invites with an error message rather
-    than 404)."""
-    return await session.get(Invitation, invitation_id)
-
-
-async def list_pending_invitations(
-    session: AsyncSession, org_id: str
-) -> list[tuple[Invitation, User]]:
-    """List non-terminal invites for an org, paired with the inviter.
-
-    Returned tuple shape mirrors :func:`list_org_members` for symmetry —
-    the dashboard's Members tab renders both as adjacent lists.
-    """
-    stmt = (
-        select(Invitation, User)
-        .join(User, User.id == Invitation.invited_by_user_id)
-        .where(
-            Invitation.org_id == org_id,
-            Invitation.accepted_at.is_(None),  # type: ignore[union-attr]
-            Invitation.revoked_at.is_(None),  # type: ignore[union-attr]
-        )
-        .order_by(Invitation.created_at)  # type: ignore[attr-defined]
-    )
-    return [(i, u) for i, u in (await session.exec(stmt)).all()]
-
-
-def _ensure_utc_aware(dt):
-    """Re-attach UTC tz to a datetime that lost it during a DB round-trip.
-
-    SQLite stores TIMESTAMP as a string with no timezone info, so a
-    value written as ``utcnow()`` (tz-aware UTC) comes back naive.
-    Comparing a naive value against a tz-aware ``utcnow()`` raises
-    ``TypeError``. The fix: treat naive DB-loaded values as already-
-    UTC and attach the tz before any comparison.
-    """
-    from datetime import timezone
-
-    if dt.tzinfo is None:
-        return dt.replace(tzinfo=timezone.utc)
-    return dt
-
-
-def _is_invitation_terminal(invitation: Invitation) -> bool:
-    """True if the invitation is already accepted, revoked, or expired."""
-    if invitation.accepted_at is not None or invitation.revoked_at is not None:
-        return True
-    return _ensure_utc_aware(invitation.expires_at) < utcnow()
-
-
-async def accept_invitation(
-    session: AsyncSession,
-    *,
-    invitation: Invitation,
-    accepting_user: User,
-) -> OrganizationMember:
-    """Atomically: validate + insert OrganizationMember + mark consumed.
-
-    Raises:
-      * :class:`InvitationExpired`         — past ``expires_at``
-      * :class:`InvitationAlreadyConsumed` — already accepted or revoked
-      * :class:`InvitationEmailMismatch`   — caller's email ≠ invited email
-
-    Idempotent on retry only in the same session — once ``accepted_at``
-    is set, a second call raises ``InvitationAlreadyConsumed``. That's
-    intentional: a double-click on the accept button shouldn't silently
-    succeed; the UI should refresh and notice the user is already a
-    member.
-
-    If the user is already a member of the org (e.g., manual add or
-    re-invite), the existing membership row is returned and the
-    invitation is still marked consumed. The role is *upgraded* (never
-    downgraded) when the invitation's role outranks the current one —
-    so an owner can promote a teammate from member → admin by sending
-    a new invite, but a member-tier re-invite of an existing admin or
-    owner is a no-op rather than a silent demotion.
-    """
-    if invitation.accepted_at is not None or invitation.revoked_at is not None:
-        raise InvitationAlreadyConsumed("invitation already accepted or revoked")
-    if _ensure_utc_aware(invitation.expires_at) < utcnow():
-        raise InvitationExpired("invitation expired")
-    if invitation.email.lower() != accepting_user.email.lower():
-        # Don't echo ``invitation.email`` back: the route forwards
-        # ``str(exc)`` as the HTTP detail, and a logged-in attacker who
-        # got hold of an invite id would otherwise be able to harvest
-        # the invitee's address. Stay generic.
-        raise InvitationEmailMismatch("invitation is for a different account")
-
-    # If already a member (re-invite of an existing teammate), reuse
-    # the existing row. Otherwise create a fresh membership.
-    member = await find_member(
-        session, org_id=invitation.org_id, user_id=accepting_user.id
-    )
-    if member is None:
-        member = OrganizationMember(
-            user_id=accepting_user.id,
-            org_id=invitation.org_id,
-            role=invitation.role,
-        )
-        session.add(member)
-    elif _ROLE_RANK.get(invitation.role, -1) > _ROLE_RANK.get(member.role, -1):
-        # Re-invite is a promotion: an admin invited as owner, or a
-        # member invited as admin. Upgrade in place. We never downgrade
-        # on the reverse — a stray member-level re-invite of an
-        # existing owner mustn't silently strip privileges.
-        member.role = invitation.role
-        session.add(member)
-    invitation.accepted_at = utcnow()
-    session.add(invitation)
-    await session.commit()
-    await session.refresh(member)
-    return member
-
-
-async def revoke_invitation(session: AsyncSession, invitation: Invitation) -> None:
-    """Mark an invitation revoked. Idempotent — already-terminal
-    invites silently no-op (the caller has already gotten the desired
-    outcome)."""
-    if invitation.accepted_at is not None or invitation.revoked_at is not None:
-        return
-    invitation.revoked_at = utcnow()
-    session.add(invitation)
-    await session.commit()
-
-
-async def send_invitation_email(
-    *,
-    invitation: Invitation,
-    org_name: str,
-    inviter_email: str,
-) -> None:
-    """Mail the invitee a clickable accept link.
-
-    Same HEXGATE_DASHBOARD_URL the verify/reset flows use, same
-    StderrEmailSender dev mode, same swap-for-real-provider story for
-    production via :func:`mailer.set_email_sender`.
-    """
-    import os
-
-    from hexgate_api.core.mailer import get_email_sender
-
-    dashboard_url = os.environ.get(
-        "HEXGATE_DASHBOARD_URL", "http://localhost:5173"
-    ).rstrip("/")
-    link = f"{dashboard_url}/invites/{invitation.id}/accept"
-    ttl_hours = max(
-        1,
-        int(
-            (_ensure_utc_aware(invitation.expires_at) - utcnow()).total_seconds() / 3600
-        ),
-    )
-
-    body = (
-        f"Hi,\n\n"
-        f'{inviter_email} invited you to join the "{org_name}" '
-        f"organisation on Hexgate as a {invitation.role}.\n\n"
-        f"Open this link to accept (you'll be prompted to sign in or sign\n"
-        f"up first if you don't have an account):\n\n"
-        f"    {link}\n\n"
-        f"The link expires in about {ttl_hours} hours. If you don't\n"
-        f"recognise the sender, just delete this email — nothing happens\n"
-        f"until you click.\n"
-    )
-    await get_email_sender().send(
-        to=invitation.email,
-        subject=f"{inviter_email} invited you to {org_name} on Hexgate",
-        body=body,
-    )
-
-
-# ---------------------------------------------------------------------------
-# First-boot seeding — the triple-default Org + User + Membership + Project
-# + agents. Cross-domain by nature; lives here until a dedicated bootstrap
-# module owns it.
-# ---------------------------------------------------------------------------
-
-
-def _announce_default_admin_credentials(email: str, password: str) -> None:
-    """Loud one-shot stderr print of the freshly-generated admin password.
-
-    Same posture as ``FileKeyStore._announce_first_run`` — operators
-    only see this once, ever; subsequent boots are silent. The password
-    is never logged again from anywhere in the codebase.
-    """
-    import sys
-
-    bar = "=" * 72
-    print(
-        f"\n{bar}\n"
-        f"FIRST-BOOT DEFAULT ADMIN CREDENTIALS\n"
-        f"   email:    {email}\n"
-        f"   password: {password}\n\n"
-        f"This is printed ONCE on first boot. Save it now — there is no\n"
-        f"second display. Sign in at the dashboard and rotate the password\n"
-        f"via your account settings as soon as you're in.\n"
-        f"\n"
-        f"Self-hosted deployments that don't want a default account at\n"
-        f"all should set HEXGATE_SEED=skip and POST /v1/auth/register\n"
-        f"to bootstrap their first user from scratch.\n"
-        f"{bar}\n",
-        file=sys.stderr,
-        flush=True,
-    )
-
-
-async def ensure_default_seed(session: AsyncSession) -> Project | None:
-    """Idempotently create the triple-default: Org + User + Membership + Project + agents.
-
-    First-boot UX for self-hosters and `make platform-api`. Every step is
-    individually idempotent so calling this on an already-seeded DB is a
-    no-op — same shape `ensure_default_project` used to have, just broader.
-
-    Returns the default Project, or ``None`` when ``HEXGATE_SEED=skip``
-    is set (production hosted deployments). When skipped, callers must
-    handle the empty-DB case explicitly — there is no implicit project.
-    """
-    if _seed_disabled():
-        return None
-
-    # Org first — Project FKs to it, so it has to exist before the project.
-    org = await session.get(Organization, DEFAULT_ORG_ID)
-    if org is None:
-        org = Organization(
-            id=DEFAULT_ORG_ID,
-            slug=DEFAULT_ORG_SLUG,
-            name=DEFAULT_ORG_NAME,
-        )
-        session.add(org)
-
-    # Default admin user. M3 Phase 3a: first boot generates a fresh
-    # random password, hashes it via FastAPI Users' PasswordHelper, and
-    # prints the plaintext to stderr ONCE for the operator to copy. On
-    # every subsequent boot the row already exists → no print, no
-    # re-hash, no behaviour change. Production deployments that don't
-    # want a default account set HEXGATE_SEED=skip and create their
-    # first user via POST /v1/auth/register instead.
-    user = await session.get(User, DEFAULT_USER_ID)
-    if user is None:
-        from fastapi_users.password import PasswordHelper
-
-        password_plain = secrets.token_urlsafe(16)
-        hashed = PasswordHelper().hash(password_plain)
-        user = User(
-            id=DEFAULT_USER_ID,
-            email=DEFAULT_USER_EMAIL,
-            hashed_password=hashed,
-            is_active=True,
-            # Default seed user is auto-verified — no email flow runs at
-            # `make platform-api`. Real registered users start unverified.
-            is_verified=True,
-            is_superuser=True,
-        )
-        session.add(user)
-        _announce_default_admin_credentials(DEFAULT_USER_EMAIL, password_plain)
-
-    # Owner membership wiring user → org. The unique constraint on
-    # (user_id, org_id) makes this safe to re-add on subsequent boots.
-    member = await session.get(OrganizationMember, DEFAULT_MEMBERSHIP_ID)
-    if member is None:
-        member = OrganizationMember(
-            id=DEFAULT_MEMBERSHIP_ID,
-            user_id=DEFAULT_USER_ID,
-            org_id=DEFAULT_ORG_ID,
-            role="owner",
-        )
-        session.add(member)
-
-    project = await session.get(Project, DEFAULT_PROJECT_ID)
-    if project is None:
-        project = Project(
-            id=DEFAULT_PROJECT_ID,
-            org_id=DEFAULT_ORG_ID,
-            name=DEFAULT_PROJECT_NAME,
-        )
-        session.add(project)
-
-    await session.commit()
-    await session.refresh(project)
-    # Always ensure seeded agents exist — idempotent, so existing projects
-    # pick up the `default` guarantee on any subsequent boot.
-    await ensure_seeded_agents(session, project.id)
-    return project
-
-
-# Back-compat alias for callers that still use the old name. New code uses
-# ``ensure_default_seed`` directly; this one-liner keeps existing imports
-# (main.py, tests) working without a renaming sweep this turn.
-ensure_default_project = ensure_default_seed
+logger = logging.getLogger("hexgate.platform.agents")
 
 
 async def ensure_seeded_agents(session: AsyncSession, project_id: str) -> None:
