+++ platform/api/tests/test_auth_oauth.py
@@ -38,7 +38,7 @@ from sqlmodel.ext.asyncio.session import AsyncSession
 from hexgate_api import main
 from hexgate_api.main import app
 from hexgate_api.models import OAuthAccount, User
-from hexgate_api.services import ensure_default_project
+from hexgate_api.bootstrap.seed import ensure_default_project
 
 
 # ---------------------------------------------------------------------------
