+++ platform/api/tests/test_invites.py
@@ -33,7 +33,7 @@ from hexgate_api.models import (
 )
 from hexgate_api.models import utcnow
 from hexgate_api.constants import ROLE_ADMIN, ROLE_MEMBER, ROLE_OWNER
-from hexgate_api.services import ensure_default_project
+from hexgate_api.bootstrap.seed import ensure_default_project
 
 
 # ---------------------------------------------------------------------------
