+++ platform/api/tests/test_orgs.py
@@ -42,7 +42,7 @@ from hexgate_api.domains.orgs.service import (
     ensure_personal_default_org,
     list_orgs_for_user,
 )
-from hexgate_api.services import ensure_default_project
+from hexgate_api.bootstrap.seed import ensure_default_project
 
 
 # ---------------------------------------------------------------------------
