+++ platform/api/tests/test_projects.py
@@ -29,7 +29,7 @@ from hexgate_api import main
 from hexgate_api.main import app
 from hexgate_api.models import OrganizationMember, User
 from hexgate_api.constants import ROLE_ADMIN, ROLE_MEMBER
-from hexgate_api.services import ensure_default_project
+from hexgate_api.bootstrap.seed import ensure_default_project
 
 
 # ---------------------------------------------------------------------------
