+++ platform/api/tests/test_tenant_isolation.py
@@ -23,12 +23,9 @@ from sqlmodel.ext.asyncio.session import AsyncSession
 
 from hexgate_api.main import app
 from hexgate_api.models import Agent, Organization, OrganizationMember, Project, User
-from hexgate_api.services import (
-    DEFAULT_PROJECT_ID,
-    DEFAULT_USER_ID,
-    ensure_default_project,
-    new_id,
-)
+from hexgate_api.core.ids import new_id
+from hexgate_api.bootstrap.seed import ensure_default_project
+from hexgate_api.constants import DEFAULT_PROJECT_ID, DEFAULT_USER_ID
 
 
 # ---------------------------------------------------------------------------
